+++ examples/demo/ui.R
@@ -6,15 +6,15 @@ header <- dashboardHeader(
 sidebar <- dashboardSidebar(
   sidebarMenu(
     menuItem("Welcome", tabName = "welcome", icon = icon("star")),
-    menuItem("Animations", tabName = "animation", icon = animateIcon("carrot", "float", color = "orange"), badgeLabel = "0.2 NEW"),
-    menuItem("Loaders", tabName = "loader", icon = animateIcon("spinner", "spin", color = "teal"), badgeLabel = "0.2 NEW"),
+    menuItem("Animations", tabName = "animation", icon = animateIcon("carrot", "float", color = "orange")),
+    menuItem("Loaders", tabName = "loader", icon = animateIcon("spinner", "spin", color = "teal")),
     menuItem("Tooltip, Popover", tabName = "poptip", icon = animateIcon("info", "rising", color = "#0275d8"), badgeLabel = "0.3 NEW"),
     menuItem("Input Control", tabName = "text", icon = icon("font")),
     menuItem("Image Display", tabName = "images", icon = icon("images")),
     menuItem("Buttons", tabName = "buttons", icon = icon("square")),
     menuItem("Progress control", tabName = "progress", icon = icon("tasks")),
     menuItem("Upload components", tabName = "upload_file", icon = icon("file-upload")),
-    menuItem("Server Components", tabName = "server_col", icon = icon("server")),
+    menuItem("Server Components", tabName = "server_col", icon = animateIcon("server"), badgeLabel = "0.4 upd"),
     menuItem("Misc", tabName = "other", icon = icon("border-all"), badgeLabel = "0.3 Updates")
   )
 )