+++ examples/demo/www/demo.js
@@ -6,6 +6,7 @@ $(function(){
     a[href="#shiny-tab-animation"],
     a[href="#shiny-tab-loader"],
     a[href="#shiny-tab-poptip"],
+    a[href="#shiny-tab-server_col"],
     a[href="#shiny-tab-other"]`);
   el.find("small").addClass('faa-pulse animated');
   el.on('click', function() {